requirements.txt
Flask==2.3.3
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
